[GraphQL/Data] Consistent Reads API
Adapt the `QueryExecutor` interface, to allow multiple queries to be
issued within one transaction that sees a consistent snapshot of the
database.

The previous API required that a query executor offer a way to run a
query that produced a single result, multiple results, and an optional
result (and provided a helper that produced one result),
asynchronously.

The new API splits this functionality into two parts:

- `QueryExecutor` which offers an asynchronous API for executing
  transactions.
- `DbConnection` which offers a way to run a query that produces one
  or more results.

When `QueryExecutor` executes a transaction, it passes it an instance
of `DbConnection`, which can then be used to run queries.

This change also gets rid of the `optional` helper, because it's no
longer required -- we can rely on Diesel's own `OptionalExtension` to
convert the output from a `result` or `first` query into an optional
query.

Finally, we also adapt all the re-usable query building
components (pagination, type filtering, etc) to accept a
`DbConnection`, instead of the `Db` to allow them to be called within
a transaction that contains other queries.

Test Plan:

This is a behaviour preserving refactor, confirmed by running all unit
and E2E tests:

```
sui-graphql-rpc$ cargo nextest run
sui-graphql-e2e-tests$ cargo nextest run -j 1 --features pg_integration
```